Move all searching code into `monument::search`
